Done resend otp

+++ paywellservices/src/main/java/com/cloudwell/paywell/services/activity/home/model/ResponseDetails.kt
@@ -6,6 +6,6 @@ class ResponseDetails {
     @SerializedName("StatusName")
     var statusName: String = ""
     @SerializedName("Status")
-    var status: Long = 0
+    var status: Int = 0
 
 }
